[ch108] Decompose Navbar component
