+++ src/components/NavbarLogo/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const NavbarLogo = () => {
+  return (
+    <div className="navbar__logo">
+      {/* TODO: Use logo instead*/}
+      <span>COBDA</span>
+    </div>
+  )
+}
+
+export default NavbarLogo
